requirements.txt
Flask==1.0.2
Werkzeug==0.15.5
flask_cors==3.0.9 
gunicorn==19.9.0
validate_email==1.3
nose==1.3.7
argparse==1.4.0
uuid==1.30
flask-swagger-ui==3.20.9

certifi>=2021.5.30
charset-normalizer>=2.0.4
filelock>=3.0.12
gdown>=3.13.0
idna>=3.2
PySocks>=1.7.1
six==1.16.0
urllib3==1.26.6
numpy>=1.19.4
scikit-image>=0.17.2
torch>=1.7.0
torchvision>=0.8.1
waitress>=1.4.4
requests>=2.24.0
scipy>=1.5.4
pymatting>=1.1.1
filetype>=1.0.7
hsh>=1.1.0
more_itertools==8.7.0
moviepy==1.0.3
Pillow==8.1.1
ffmpeg-python
opencv-python
tensorflow
tf_slim